[OOO-core] Further sanitising of the dispatch deadlock watchdog. The dispatch watchdog could cause a flush_pipeline just as an x86 instruction was half-way committed and already updated *some* architectural state. For example the "ret" instruction updates the RSP register and then branches. Flushing only the branch part leads to the RSP being modified again on re-execution of the "ret" instruction.
Fix:
Calm down the watchdog, when the core is still making progress, e.g. when there
are still uops ready for commit.

The general pipeline flush logic should be changed to record the requested flush
and flush later on x86 instruction boundaries. For now I've added a debug printout
that notifies the user of such issues (loglevel >= 1) to see if that actually
happens.