Fix ptlstats' -subtract for general use
The -subtract option of ptlstats does not appear to work in all
cases, in particular the case where you might want to view all
of the stats while subtracting some previous snapshot, e.g.
% ptlstats -subtract 0 -snapshot 1 stats-file

Surprisingly, that didn't work.  This fix makes it work.

It appeared that getdelta() was only used in certain cases but not the
general case.  This patch creates what I consider to be expected
behavior (based on the documentation).