Curtis Dunham: Fix broken 32bit syscall semantics.
Commit 9687405f1d10183af7e910deac79e925794b7175 in the official PTLsim
repo broke my PTLsim.

The reason is that the inline assembly must save the ebp register and load
eax with the syscall number while taking special care to deal with the
stack pointer correctly.  The previous version loaded eax from a static
variable that had a fixed offset, so it wasn't esp-relative.  Then when it
was changed to a local variable, it became esp-relative and the push/pop
pair broke the addressing while trying to save/restore ebp.

This change removes the use of push and pop for saving ebp.

Thanks Curtis!

SD: 64 bit should not have the issue, as the value for %ebp is passed
    in through another register.