Prevent livelock from unaligned stores causing SMC
[SMC = self-modifying code]
If an unaligned store caused the SMC detection, the pipeline
would flush between the two aligned store uops, causing a
livelock.  Changed SMC handling in the pipeline so that
the current instruction completes (allows EOM uop to commit
before flushing the pipeline).

More detail...
There is a comment in ooopipe.cpp ThreadContext::flush_pipeline:
  // SD: I wonder if flush_pipeline should really be able to flush halfway
  // through a partially committed x86 instruction. This is dangerous,
  // especially if the instruction has already partially updated
  // architectural state.
Without this patch, this 'dangerous' situation breaks correctness
for self-modifying code and causes the above livelock.

I'm not sure this patch is 100% correct but it does fix a testcase
I ran into.